feat: product card
