+++ client/src/type/product.ts
@@ -1,0 +1,8 @@
+export type ProductData = {
+  name: string;
+  tags: string[];
+  thumbnail: string;
+  subThumbnail: string[];
+  detail: string;
+  price: string;
+};
